Refactor packetbeat build logic
