+++ packetbeat/Makefile
@@ -1,15 +1,11 @@
-BEAT_NAME?=packetbeat
-BEAT_TITLE?=Packetbeat
-SYSTEM_TESTS?=true
-TEST_ENVIRONMENT=false
-ES_BEATS?=..
-EXCLUDE_COMMON_UPDATE_TARGET=true
+#
+# Includes
+#
+include ../dev-tools/make/oss.mk
 
-include ${ES_BEATS}/libbeat/scripts/Makefile
-
-.PHONY: update
-update: mage
-	mage update
+#
+# Targets
+#
 
 .PHONY: benchmark
 benchmark:
@@ -17,4 +13,4 @@ benchmark:
 
 .PHONY: create-tcp-protocol
 create-tcp-protocol: python-env
-	@${PYTHON_ENV}/bin/python ${ES_BEATS}/packetbeat/scripts/create_tcp_protocol.py
+	@${PYTHON_VE_DIR}/bin/python ${ES_BEATS}/packetbeat/scripts/create_tcp_protocol.py
